+++ src/index.ts
@@ -1,3 +1,5 @@
+export { ReactScrimProvider } from "./components/provider.js";
 export { Scrim } from "./components/scrim.js";
+export { useReactScrim } from "./components/use_scrim.js";
 
-export type { ScrimProps, ScrimSignal } from "./components/type.js";
+export type { ReactScrimProviderProps, ScrimPhase, ScrimProps, ScrimState } from "./components/type.js";